ID##188873729: Refactor code to ensure modularity, cohesion, and loose coupling
